fix: 캐릭터 현황 코드 개선
