+++ client/src/App.tsx
@@ -2,16 +2,45 @@ import React, { useRef } from "react";
 import styled, { ThemeProvider } from "styled-components";
 import "./App.css";
 import AccountBook from "./components/AccountBook";
-import Characters from "./components/Characters";
 import ContentsStatus from "./components/ContentsStatus";
-import { Character } from "./components/character/FCharacter";
-import { CharacterStatus } from "./components/character/FCharacterStatus";
+import { CharacterStatus } from "./components/CharacterStatus";
 import FixedMenu from "./components/FixedMenu";
 import Footer from "./components/Footer";
 import HeaderMenu from "./components/HeaderMenu";
 import Monitoring from "./components/Monitoring";
 import { RootState, useAppSelector } from "./redux/store";
 import { theme } from "./style/theme";
+
+const App = () => {
+  const accountBookRef = useRef<HTMLDivElement>(null);
+  const characterRef = useRef<HTMLDivElement>(null);
+  const monitoringRef = useRef<HTMLDivElement>(null);
+  const contentsStatusRef = useRef<HTMLDivElement>(null);
+  const themeType = useAppSelector(
+    (state: RootState) => state.themeReducer.themeType
+  );
+
+  return (
+    <ThemeProvider theme={theme[themeType]}>
+      <Wrap>
+        <HeaderMenu />
+        <section>
+          <Monitoring ref={monitoringRef} />
+          <CharacterStatus ref={characterRef} />
+          <ContentsStatus ref={contentsStatusRef} />
+          <AccountBook ref={accountBookRef} />
+        </section>
+        <FixedMenu
+          accountBookRef={accountBookRef}
+          characterRef={characterRef}
+          monitoringRef={monitoringRef}
+          contentsStatusRef={contentsStatusRef}
+        />
+        <Footer />
+      </Wrap>
+    </ThemeProvider>
+  );
+};
 
 const Wrap = styled.div`
   background-color: ${({ theme }) => theme.color.backgroundColor};
@@ -27,37 +56,5 @@ const Wrap = styled.div`
     }
   }
 `;
-
-function App() {
-  const accountBookRef = useRef<HTMLDivElement>(null);
-  const characterRef = useRef<HTMLDivElement>(null);
-  const monitoringRef = useRef<HTMLDivElement>(null);
-  const contentsStatusRef = useRef<HTMLDivElement>(null);
-  const themeType = useAppSelector(
-    (state: RootState) => state.themeReducer.themeType
-  );
-
-  return (
-    <ThemeProvider theme={theme[themeType]}>
-      <Wrap>
-        <HeaderMenu />
-        <section>
-          <CharacterStatus />
-          <Monitoring ref={monitoringRef} />
-          <Characters ref={characterRef} />
-          <ContentsStatus ref={contentsStatusRef} />
-          <AccountBook ref={accountBookRef} />
-        </section>
-        <FixedMenu
-          accountBookRef={accountBookRef}
-          characterRef={characterRef}
-          monitoringRef={monitoringRef}
-          contentsStatusRef={contentsStatusRef}
-        />
-        <Footer />
-      </Wrap>
-    </ThemeProvider>
-  );
-}
 
 export default App;
